PyTorchAnnual_SoilOutlier/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html

certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.1
cycler==0.10.0
dataclasses==0.6
Flask==2.0.1
Flask-API==3.0.post1
future==0.18.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
numpy==1.21.2
pandas==1.3.4
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
torch==1.7.1+cpu
torchvision==0.8.2+cpu
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.1
python-dotenv==0.19.1
